App manager UI updated and cleaned up.
