+++ src/assets/app-definition.ts
@@ -3,7 +3,6 @@
  * When these functions are called, then the actual import will happen.
  */
 export const appImportMap = {
-  hello: () => import('hello/HelloApp' as any),
   simpleMenu: () => import('simpleMenu/SimpleMenuApp' as any),
   simplePanel: () => import('simplePanel/SimplePanelApp' as any),
 }
