+++ src/assets/apps-dev.json
@@ -1,10 +1,5 @@
 [
   {
-    "name": "hello",
-    "url": "http://localhost:2222/remoteEntry.js",
-    "entryPoint": "HelloApp"
-  },
-  {
     "name": "simpleMenu",
     "url": "http://localhost:3333/remoteEntry.js",
     "entryPoint": "SimpleMenuApp"
